+++ src/misc/akarin_timer.cpp
@@ -7,9 +7,9 @@
 
 // TODO :: Add a hierarchial view of markers
 
-std::unordered_map<const char *, double> chrono_timer_marks;
+std::unordered_map<const char *, float> chrono_timer_marks;
 
-double get_current_time() noexcept;
+float get_current_time() noexcept;
 
 void AkarinTimer::mark_begin(
     const char *p_mark) noexcept
@@ -23,7 +23,7 @@ void AkarinTimer::mark_end(
     auto find_iter = std::find_if(
         chrono_timer_marks.cbegin(),
         chrono_timer_marks.cend(),
-        [p_mark](const std::pair<const char *, double> &p_timer_iter) -> bool {
+        [p_mark](const std::pair<const char *, float> &p_timer_iter) -> bool {
             return strcmp(p_mark, p_timer_iter.first) == 0;
         });
     if (find_iter == chrono_timer_marks.cend())
@@ -37,19 +37,19 @@ void AkarinTimer::mark_end(
     std::cout << "timer_mark name:" << p_mark << " time:" << current_time - chrono_timer_marks[p_mark] << "\n";
 };
 
-double get_current_time() noexcept
+float get_current_time() noexcept
 {
     static const auto start_time = std::chrono::high_resolution_clock::now();
     auto end = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> elapsed_seconds = end - start_time;
+    std::chrono::duration<float> elapsed_seconds = end - start_time;
     return elapsed_seconds.count();
 };
 
-double AkarinTimer::get_delta_time() noexcept
+float AkarinTimer::get_delta_time() noexcept
 {
     static auto start_time = std::chrono::high_resolution_clock::now();
     auto end = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> elapsed_seconds = end - start_time;
+    std::chrono::duration<float> elapsed_seconds = end - start_time;
     start_time = end;
     return elapsed_seconds.count();
 };
